requirements.txt
requests
python-dotenv
matplotlib
pandas
numpy
